requirements.txt
nbtlib
Pillow
